convert gaussian bounds to u32

+++ crates/engine/renderers/tataku-wgpu/shaders/gaussian_blur.wgsl
@@ -3,10 +3,10 @@
 
 struct Settings {
     filter_size: u32,
-    x: f32,
-    y: f32,
-    width: f32,
-    height: f32,
+    x: u32,
+    y: u32,
+    width: u32,
+    height: u32,
 };
 
 struct Orientation {
